feat(backend): updates e creates de comunidades
implementação do update de comunidades, da criação de banners para comunidades e sua atualização na
camada de service

+++ GameLink-API/src/main/java/com/gamelink/gamelinkapi/repositories/images/CloudinaryRepository.java
@@ -25,9 +25,7 @@ public class CloudinaryRepository {
                     .uploader()
                     .upload(fileConverted, ObjectUtils.emptyMap());
 
-            if (fileConverted != null) {
-                fileConverted.delete();
-            }
+            if (fileConverted != null) fileConverted.delete();
 
             return ImageModel.builder()
                     .filename(file.getName())
